requirements.txt
beautifulsoup4==4.4.1
pytest==2.7.3
pytest-selenium
requests==2.8.0
